refactor(server): added startUnixSocket helper
