+++ lib/Server.js
@@ -6,7 +6,6 @@
   func-names
 */
 const fs = require('fs');
-const net = require('net');
 const path = require('path');
 const tls = require('tls');
 const url = require('url');
@@ -36,6 +35,7 @@ const routes = require('./utils/routes');
 const getSocketServerImplementation = require('./utils/getSocketServerImplementation');
 const handleStdin = require('./utils/handleStdin');
 const tryParseInt = require('./utils/tryParseInt');
+const startUnixSocket = require('./utils/startUnixSocket');
 const schema = require('./options.json');
 
 // Workaround for node ^8.6.0, ^9.0.0
@@ -776,55 +776,7 @@ class Server {
       // set this so that status helper can identify how the project is being run correctly
       this.options.socket = socket;
 
-      const chmodSocket = (done) => {
-        // chmod 666 (rw rw rw) - octal
-        const READ_WRITE = 438;
-        fs.chmod(socket, READ_WRITE, done);
-      };
-
-      const startSocket = () => {
-        this.listeningApp.on('error', (err) => {
-          userCallback(err);
-        });
-
-        this.listeningApp.listen(socket, hostname, (err) => {
-          if (err) {
-            fullCallback(err);
-          } else {
-            chmodSocket(fullCallback);
-          }
-        });
-      };
-
-      fs.access(socket, fs.constants.F_OK, (err) => {
-        if (err) {
-          // file does not exist
-          startSocket();
-        } else {
-          // file exists
-
-          const clientSocket = new net.Socket();
-
-          clientSocket.on('error', (err) => {
-            if (err.code === 'ECONNREFUSED' || err.code === 'ENOTSOCK') {
-              // No other server listening on this socket so it can be safely removed
-              fs.unlinkSync(socket);
-
-              startSocket();
-            }
-          });
-
-          clientSocket.connect({ path: socket }, () => {
-            // if a client socket successfully connects to the given socket path,
-            // it means that the socket is in use
-            const err = new Error('This socket is already used');
-            clientSocket.destroy();
-            // do not call onListening or the setup method, since the server
-            // cannot start listening on a used socket
-            userCallback(err);
-          });
-        }
-      });
+      startUnixSocket(this.listeningApp, socket, fullCallback, userCallback);
     } else {
       this.listeningApp.listen(port, hostname, fullCallback);
     }
